pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.69",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.121.0",
    "langchain>=1.0.5",
    "langchain-aws>=1.0.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-distro>=0.59b0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-langchain>=0.48.0",
    "opentelemetry-sdk>=1.38.0",
    "pydantic-settings>=2.11.0",
]

[tool.mypy]
disallow_untyped_defs = false

[tool.ruff]
target-version = "py313"
exclude = [".venv"]
line-length = 119

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    # "ARG001", # unused arguments in functions
]


ignore = [
    "B008", # do not perform function calls in argument defaults
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["F821"]
